refactor apply.go
