+++ nodes/vr_node.go
@@ -50,8 +50,8 @@ func NewVRNode(n NodeOverwrites, creds *Credentials, scrapliPlatformName string)
 
 // LinkApplyMode keeps vrnetlab-backed VM nodes on the conservative recreate
 // path for apply link changes.
-func (*VRNode) LinkApplyMode() LinkApplyMode {
-	return LinkApplyModeRecreate
+func (vr *VRNode) LinkApplyMode(ctx context.Context) LinkApplyMode {
+	return vr.ImageLinkApplyMode(ctx, LinkApplyModeRecreate)
 }
 
 // Init stub function.
